Refactor trantor package
