+++ pkg/trantor/testing/smr_property_test.go
@@ -1,4 +1,4 @@
-package trantor
+package testing
 
 import (
 	"bytes"
